Rename `managedDatabase` -> `appPostgres`
