+++ apps/infra/src/k8s/postgres.ts
@@ -1,7 +1,7 @@
 import * as k8s from '@pulumi/kubernetes';
 import { type core } from '@pulumi/kubernetes/types/input';
 import { all, secret, type Input } from '@pulumi/pulumi';
-import { appDatabase, appPg } from '../vultr/managedDatabase';
+import { appDatabase, appPg } from '../vultr/appPostgres';
 import { provider } from './provider';
 
 export const cloudNativePg = new k8s.helm.v3.Release('cloud-native-pg', {
@@ -83,7 +83,7 @@ export const getConnectionDetails = (resource: k8s.apiextensions.CustomResource)
   };
 };
 
-// Connection URI for the managed database.
+// Connection URI for appPg.
 // sslmode=no-verify: node-postgres can reject Vultr's cert chain, so we encrypt but skip CA
 // verification (same posture as the commented-out backend config in serviceDefinitions.ts).
 // The password is URL-encoded in case it contains reserved characters (@ : / etc).
